response data issue fixed

+++ app/src/main/java/com/manish/stockapp/model/StockDetailsModel.kt
@@ -1,11 +1,11 @@
 package com.manish.stockapp.model
 
 data class StockDetailsModel(val sid:String,
-                             val price:Long,
-                             val close:Long,
-                             val change:Long,
-                             val high:Long,
-                             val low:Long,
+                             val price:Float,
+                             val close:Float,
+                             val change:Float,
+                             val high:Float,
+                             val low:Float,
                              val volume:Long,
                              val date:String,
